Add new charts
