+++ src/client/components/Sec/LanguagesParticipants/style.scss
@@ -1,20 +1,18 @@
 @import '~Colors';
 
 .wrapper {
-  display: flex;
-  margin-bottom: 20px;
   flex: 1;
+  display: flex;
   flex-direction: column;
+  margin-bottom: 20px;
   background-color: $bg_dark;
   overflow-y: scroll;
-  grid-column: 1 / 4;
-  grid-row: 1 / 3;
-  height: 50%;
 }
 
 .header {
   align-items: center;
   padding: 13px;
+  margin: 0;
   border-bottom: 1px solid $MINE_SHAFT;
 }
 
@@ -25,4 +23,12 @@
   line-height: 1.71;
   text-align: left;
   color: $WHITE;
+}
+
+.doughnut {
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0 10px;
 }
