+++ src/client/components/Sec/LanguagesTasks/style.scss
@@ -1,19 +1,17 @@
 @import '~Colors';
 
 .wrapper {
-  display: flex;
   flex: 1;
+  display: flex;
   flex-direction: column;
   background-color: $bg_dark;
   overflow-y: scroll;
-  grid-column: 1 / 4;
-  grid-row: 1 / 3;
-  height: 45%;
 }
 
 .header {
   align-items: center;
   padding: 13px;
+  margin: 0;
   border-bottom: 1px solid $MINE_SHAFT;
 }
 
@@ -24,4 +22,12 @@
   line-height: 1.71;
   text-align: left;
   color: $WHITE;
+}
+
+.horBar {
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0 10px;
 }
