--- src/client/components/Sec/shared/style.scss
@@ -1,1 +1,0 @@
-@import '~Colors';
